moved file cache fix function to CacheManager class
